# api/openapi.gen.yaml
openapi: 3.0.3
info:
  title: ruby-core API
  version: 0.1.0
  description: |
    ruby-core's synchronous HTTP read plane — the spec-first, self-documenting read
    API that every domain inherits (ROADMAP-0012). Authored as per-domain fragments
    and bundled into a single served document; the bundle is the source of truth for
    the generated Go server/client and the Python client (ADR-0041).

    Authentication is defense-in-depth (ADR-0040): Traefik edge auth + mTLS is the
    primary gate; the `bearerAuth` scheme below is an additional in-app layer. All
    errors are RFC 9457 Problem Details.
  contact:
    name: ruby-core
servers:
  - url: /v1
    description: Version 1. The API is versioned by URL path; breaking changes require a new path version.
security:
  - bearerAuth: []
tags:
  - name: meta
    description: Service metadata and liveness of the API surface itself.
  - name: calendar
    description: Read access to the family calendar mirror (ROADMAP-0012).
paths:
  /ping:
    get:
      operationId: ping
      tags:
        - meta
      summary: Liveness of the authenticated API surface
      description: |
        Returns a small payload confirming the API is reachable and the caller's bearer
        token was accepted. This is the placeholder endpoint that establishes the read
        platform; domain endpoints (calendar, directory, childcare) are added against it
        in later slices. Distinct from the unauthenticated `/health` probe, which lives
        outside the versioned, generated surface for Traefik and Uptime Kuma.
      responses:
        '200':
          description: The API is reachable and the bearer token was accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Always `ok` when the surface is healthy.
                  service:
                    type: string
                    description: The service name reporting liveness.
                required:
                  - status
                  - service
              example:
                status: ok
                service: api
        '401':
          description: Missing or invalid bearer token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
        default:
          description: Unexpected error, as an RFC 9457 Problem Details object.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred.
  /calendar/events:
    get:
      operationId: listCalendarEvents
      tags:
        - calendar
      summary: List calendar event instances in a date range
      description: |
        Returns a flat, sorted list of calendar event instances whose time overlaps the
        requested `[start, end)` window. Single events are returned directly; recurring
        series are expanded timezone-aware (EXDATEs and cancelled occurrences subtracted,
        modified occurrences applied). The window is bounded — a request that would expand
        an excessive range is rejected with a 400 Problem (ADR-0042).
      parameters:
        - name: start
          in: query
          required: true
          description: Inclusive start of the window, as an RFC 3339 timestamp.
          schema:
            type: string
            format: date-time
          example: '2026-06-01T00:00:00Z'
        - name: end
          in: query
          required: true
          description: Exclusive end of the window, as an RFC 3339 timestamp. Must be after start and within the maximum window.
          schema:
            type: string
            format: date-time
          example: '2026-07-01T00:00:00Z'
      responses:
        '200':
          description: The expanded, sorted instances in the window.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarInstance'
              example:
                - google_event_id: abc123
                  summary: Dentist
                  start: '2026-06-26T13:00:00Z'
                  end: '2026-06-26T14:00:00Z'
                  all_day: false
                  status: confirmed
                  location: 123 Main St
                  description: Cleaning
                  subjects: []
        '400':
          description: Invalid range (end not after start) or the window exceeds the maximum.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Bad Request
                status: 400
                detail: The requested date range exceeds the maximum allowed window.
        '401':
          description: Missing or invalid bearer token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
        default:
          description: Unexpected error, as an RFC 9457 Problem Details object.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Vault-issued service bearer token, enforced in-app as defense-in-depth behind
        the primary Traefik edge gate and mTLS (ADR-0040). LAN-only, service-to-service.
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457 Problem Details. The single error shape returned by every operation in
        this API (ADR-0041). Domain-specific context is carried in extension members rather
        than by string-munging `detail`.
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference identifying the problem type. Defaults to `about:blank` when no specific type applies.
        title:
          type: string
          description: A short, human-readable, stable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code, duplicated here for clients that read the body.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence (typically the request path).
      required:
        - title
        - status
      example:
        type: about:blank
        title: Bad Request
        status: 400
        detail: The requested date range exceeds the maximum allowed window.
        instance: /v1/calendar/events
    CalendarInstance:
      type: object
      description: |
        One concrete occurrence of a calendar event in the requested window. Recurring
        series are expanded into instances server-side, timezone-aware. Times are RFC 3339
        UTC instants; for all-day events `all_day` is true and `end` is EXCLUSIVE (a
        one-day event on the 26th ends on the 27th — surfaced, not hidden; see ADR-0042).
        `subjects` and `childcare` are resolved from the local overlay (populated in the
        household-overlay slice).
      properties:
        google_event_id:
          type: string
          description: The Google event id of the master (or override child) this instance came from.
        summary:
          type: string
          description: The event title.
        start:
          type: string
          format: date-time
          description: Instance start as an RFC 3339 UTC instant.
        end:
          type: string
          format: date-time
          description: Instance end as an RFC 3339 UTC instant (exclusive for all-day events).
        all_day:
          type: boolean
          description: Whether this is an all-day event (date-only in Google).
        status:
          type: string
          description: Google event status — confirmed, tentative, or cancelled.
        location:
          type: string
          description: Free-text location, when set.
        description:
          type: string
          description: Free-text description, when set.
        subjects:
          type: array
          items:
            type: string
          description: Resolved subject person ids (the "FOR" lane) — local overlay, never written to Google.
        childcare:
          type: string
          description: Resolved childcare provider id for this event; omitted when none — local overlay.
      required:
        - google_event_id
        - start
        - end
        - all_day
        - status
      example:
        google_event_id: abc123
        summary: Dentist
        start: '2026-06-26T13:00:00Z'
        end: '2026-06-26T14:00:00Z'
        all_day: false
        status: confirmed
        location: 123 Main St
        description: Cleaning
        subjects: []